Create one plot per mutant line and plot alongside wt
To create one plot per mutant line with the wt data plotted alongside a new column is created
in the dataframe to specify each plot. The wt data is then duplicated for each mutant line and
assigned to each plot via this new column. This allows me to specify plot via the col function
in FacetGrid. Further, another column in the dataframe is created to specify mutant or wt data
so that the seaborn hue function can be used to have consistent colours across plots for wt and
mutant line data